move all unit testing helpers to test_helpers module
