+++ crates/api_server/src/survey.rs
@@ -163,102 +163,11 @@ async fn get_survey_from_db(db: &Storage, survey_id: i32) -> anyhow::Result<Surv
 
 #[cfg(test)]
 mod tests {
-    use std::panic;
-
     use crate::db::models::SurveyQuestions;
 
     use super::*;
-    use diesel::{sql_query, Connection, PgConnection, RunQueryDsl};
-    use jsonwebtoken::EncodingKey;
+    use crate::test_helpers::*;
     use rocket::local::blocking::Client;
-
-    fn create_db_for_tests() -> String {
-        let db_name = format!("survey_app_test_{}", uuid::Uuid::new_v4()).replace("-", "_");
-        let mut conn = PgConnection::establish("postgres://vscode:notsecure@db/survey_app")
-            .expect("Failed to connect to database");
-        sql_query(format!("CREATE DATABASE {db_name}"))
-            .execute(&mut conn)
-            .expect("Failed to create test database");
-        db_name.to_string()
-    }
-
-    fn drop_test_db(db_name: String) {
-        let mut conn = PgConnection::establish("postgres://vscode:notsecure@db/survey_app")
-            .expect("Failed to connect to database");
-        sql_query(format!("DROP DATABASE IF EXISTS {db_name}"))
-            .execute(&mut conn)
-            .expect("Failed to drop test database");
-    }
-
-    fn create_test_user(client: &Client) -> String {
-        let response = client
-            .post(uri!("/api", crate::user::register_user))
-            .header(rocket::http::ContentType::JSON)
-            .body(r#"{"username": "test", "password": "test"}"#)
-            .dispatch();
-        format!(
-            "Bearer {}",
-            response
-                .into_json::<crate::user::UserToken>()
-                .unwrap()
-                .token
-        )
-    }
-
-    fn make_jwt(client: &Client, user_id: i32) -> String {
-        let key =
-            EncodingKey::from_secret(client.rocket().config().secret_key.to_string().as_bytes());
-        let claims = Claims::new(user_id);
-        let token = jsonwebtoken::encode(&jsonwebtoken::Header::default(), &claims, &key).unwrap();
-        "Bearer ".to_string() + &token
-    }
-
-    fn run_test_with_db<T>(test: T) -> ()
-    where
-        T: FnOnce(&String) -> () + panic::UnwindSafe,
-    {
-        let db_name = create_db_for_tests();
-        let result = panic::catch_unwind(|| test(&db_name));
-        drop_test_db(db_name);
-        assert!(result.is_ok())
-    }
-
-    fn test_rocket(db_name: &String) -> rocket::Rocket<rocket::Build> {
-        let rocket = crate::rocket();
-        let config = rocket
-            .figment()
-            .clone()
-            .merge((
-                "databases.survey_app_test.url",
-                format!("postgres://vscode:notsecure@db/{}", db_name),
-            ))
-            .merge(("databases.survey_app_test.pool_size", 1));
-        return rocket.configure(config);
-    }
-
-    fn create_survey(client: &Client, token: &String) -> i32 {
-        let response = client
-            .post(uri!("/api", create_survey))
-            .header(rocket::http::ContentType::JSON)
-            .header(rocket::http::Header::new("Authorization", token.clone()))
-            .dispatch();
-        response.into_json::<Survey>().unwrap().id
-    }
-
-    fn publish_survey(client: &Client, token: &String, survey_id: i32) {
-        client
-            .patch(uri!("/api", edit_survey(survey_id)).to_string())
-            .header(rocket::http::ContentType::JSON)
-            .header(rocket::http::Header::new("Authorization", token.clone()))
-            .body(
-                serde_json::to_vec(&SurveyPatch {
-                    published: Some(true),
-                    ..Default::default()
-                })
-                .unwrap(),
-            )
-            .dispatch();
-    }
 
     #[test]
     fn test_create_survey() {
@@ -282,7 +191,7 @@ mod tests {
             let client = Client::tracked(test_rocket(db_name)).expect("valid rocket instance");
 
             let token = create_test_user(&client);
-            let survey_id = create_survey(&client, &token);
+            let survey_id = make_survey(&client, &token);
 
             let response = client
                 .get(uri!("/api", get_survey(survey_id)).to_string())
@@ -300,7 +209,7 @@ mod tests {
             let client = Client::tracked(test_rocket(db_name)).expect("valid rocket instance");
 
             let token = create_test_user(&client);
-            let survey_id = create_survey(&client, &token);
+            let survey_id = make_survey(&client, &token);
             publish_survey(&client, &token, survey_id);
 
             let response = client
@@ -318,7 +227,7 @@ mod tests {
             let client = Client::tracked(test_rocket(db_name)).expect("valid rocket instance");
 
             let token = create_test_user(&client);
-            let survey_id = create_survey(&client, &token);
+            let survey_id = make_survey(&client, &token);
             let token = make_jwt(&client, 58008);
 
             let response = client
@@ -337,7 +246,7 @@ mod tests {
             let client = Client::tracked(test_rocket(db_name)).expect("valid rocket instance");
 
             let token = create_test_user(&client);
-            let survey_id = create_survey(&client, &token);
+            let survey_id = make_survey(&client, &token);
 
             let response = client
                 .get(uri!("/api", get_survey(survey_id)).to_string())
@@ -354,7 +263,7 @@ mod tests {
             let client = Client::tracked(test_rocket(db_name)).expect("valid rocket instance");
 
             let token = create_test_user(&client);
-            let survey_id = create_survey(&client, &token);
+            let survey_id = make_survey(&client, &token);
             publish_survey(&client, &token, survey_id);
 
             let token = make_jwt(&client, 58008);
@@ -375,7 +284,7 @@ mod tests {
             let client = Client::tracked(test_rocket(db_name)).expect("valid rocket instance");
 
             let token = create_test_user(&client);
-            let survey_id = create_survey(&client, &token);
+            let survey_id = make_survey(&client, &token);
 
             let response = client
                 .patch(uri!("/api", edit_survey(survey_id)).to_string())
@@ -415,7 +324,7 @@ mod tests {
             let client = Client::tracked(test_rocket(db_name)).expect("valid rocket instance");
 
             let token = create_test_user(&client);
-            let survey_id = create_survey(&client, &token);
+            let survey_id = make_survey(&client, &token);
             publish_survey(&client, &token, survey_id);
 
             let response = client
@@ -441,7 +350,7 @@ mod tests {
             let client = Client::tracked(test_rocket(db_name)).expect("valid rocket instance");
 
             let token = create_test_user(&client);
-            let survey_id = create_survey(&client, &token);
+            let survey_id = make_survey(&client, &token);
             publish_survey(&client, &token, survey_id);
 
             let response = client
@@ -467,7 +376,7 @@ mod tests {
             let client = Client::tracked(test_rocket(db_name)).expect("valid rocket instance");
 
             let token = create_test_user(&client);
-            let survey_id = create_survey(&client, &token);
+            let survey_id = make_survey(&client, &token);
             publish_survey(&client, &token, survey_id);
 
             let token = make_jwt(&client, 58008);
